Improve fish completions
Refactored some parts of the fish completions such that they now do
not use `cat` anymore. Instead the "description search" now uses
the `bartib last` command to find the descriptions. In addition
to that, I also made it such that the completion for the `continue`
command now shows to which project and description the number belongs.